refactor: 조건문 담는 객체 위치 리팩토링
