+++ src/main/java/persistence/sql/dml/query/builder/SelectQueryBuilder.java
@@ -3,7 +3,7 @@ package persistence.sql.dml.query.builder;
 import persistence.sql.dml.query.clause.ColumnClause;
 import persistence.sql.dml.query.clause.WhereClause;
 import persistence.sql.entity.EntityMappingTable;
-import persistence.sql.entity.model.Criterias;
+import persistence.sql.entity.conditional.Criterias;
 
 public class SelectQueryBuilder {
     private static final String FORMAT = "SELECT %s FROM %s %s";
