+++ src/main/java/persistence/sql/entity/conditional/Criterias.java
@@ -1,4 +1,4 @@
-package persistence.sql.entity.model;
+package persistence.sql.entity.conditional;
 
 import java.util.Collections;
 import java.util.List;
